+++ plugins/actix-web/src/lib.rs
@@ -6,7 +6,9 @@ extern crate actix_web3 as actix_web;
 pub mod web;
 
 pub use self::web::{Resource, Route, Scope};
-pub use paperclip_macros::{api_v2_errors, api_v2_operation, Apiv2Schema, Apiv2Security};
+pub use paperclip_macros::{
+    api_v2_errors, api_v2_operation, delete, get, post, put, Apiv2Schema, Apiv2Security,
+};
 
 use self::web::{RouteWrapper, ServiceConfig};
 use actix_service::ServiceFactory;